requirements.txt
streamlit>=1.34
openai-whisper
torch           # Whisper zależy od PyTorch
